web3 cleanup
