--- src/lib/swapper/swappers/utils/abi/erc20Allowance-abi.ts
@@ -1,27 +1,0 @@
-import type { AbiItem } from 'web3-utils'
-
-export const erc20AllowanceAbi: AbiItem[] = [
-  {
-    constant: true,
-    inputs: [
-      {
-        name: '_owner',
-        type: 'address',
-      },
-      {
-        name: '_spender',
-        type: 'address',
-      },
-    ],
-    name: 'allowance',
-    outputs: [
-      {
-        name: '',
-        type: 'uint256',
-      },
-    ],
-    payable: false,
-    stateMutability: 'view',
-    type: 'function',
-  },
-]
